fix: fix bug

+++ dongjun-platform/src/main/java/com/gdut/dongjun/web/TemperatureModuleController.java
@@ -2,8 +2,6 @@ package com.gdut.dongjun.web;
 
 import java.util.List;
 
-import javax.servlet.http.HttpSession;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,15 +11,10 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.gdut.dongjun.HitchConst;
 import com.gdut.dongjun.domain.model.ResponseMessage;
 import com.gdut.dongjun.domain.po.DataMonitorSubmodule;
+import com.gdut.dongjun.domain.po.ElectronicModule;
 import com.gdut.dongjun.domain.po.TemperatureModule;
-import com.gdut.dongjun.domain.po.User;
-import com.gdut.dongjun.dto.TemperatureMeasureHitchEventDTO;
-import com.gdut.dongjun.service.UserService;
-import com.gdut.dongjun.service.device.DataMonitorService;
 import com.gdut.dongjun.service.device.DataMonitorSubmoduleService;
 import com.gdut.dongjun.service.device.TemperatureModuleService;
-import com.gdut.dongjun.service.device.event.ModuleHitchEventService;
-import com.gdut.dongjun.service.device.event.TemperatureMeasureHitchEventService;
 import com.gdut.dongjun.util.MyBatisMapUtil;
 import com.gdut.dongjun.util.UUIDUtil;
 
@@ -31,14 +24,6 @@ public class TemperatureModuleController {
 
 	@Autowired
 	private TemperatureModuleService moduleService;
-	@Autowired
-	private TemperatureMeasureHitchEventService eventService;
-	@Autowired
-	private ModuleHitchEventService moduleHitchService;
-	@Autowired
-	private UserService userService;
-	@Autowired
-	private DataMonitorService monitorService;
 	@Autowired
 	private DataMonitorSubmoduleService submoduleService;
 	
@@ -52,6 +37,12 @@ public class TemperatureModuleController {
 	@Transactional
 	public ResponseMessage edit(TemperatureModule module, String monitorId) {
 		if (null == module.getId() || "".equals(module.getId())) {
+			
+			List<TemperatureModule> modules = moduleService
+					.selectByParameters(MyBatisMapUtil.warp("device_number", module.getDeviceNumber()));
+			if (0 != modules.size()) {
+				return ResponseMessage.warning("该设备地址已经被占用");
+			}
 			module.setId(UUIDUtil.getUUID());
 			if (0 == moduleService.updateByPrimaryKey(module)) {
 				return ResponseMessage.warning("操作失败");
@@ -66,6 +57,11 @@ public class TemperatureModuleController {
 				return ResponseMessage.warning("操作失败");
 			}
 		} else {
+			List<TemperatureModule> modules = moduleService
+					.selectByParameters(MyBatisMapUtil.warp("device_number", module.getDeviceNumber()));
+			if (0 != modules.size() && !modules.get(0).getId().equals(module.getId())) {
+				return ResponseMessage.warning("该设备地址已经被占用");
+			}
 			if (0 == moduleService.updateByPrimaryKeySelective(module)) {
 				return ResponseMessage.warning("操作失败");
 			}
@@ -86,7 +82,11 @@ public class TemperatureModuleController {
 		if (!moduleService.deleteByPrimaryKey(id)) {
 			return ResponseMessage.warning("操作失败");
 		}
-		DataMonitorSubmodule submodule = submoduleService.selectByParameters(MyBatisMapUtil.warp("module_id", id)).get(0);
+		List<DataMonitorSubmodule> submodules = submoduleService.selectByParameters(MyBatisMapUtil.warp("module_id", id));
+		if (0 == submodules.size()) {
+			return ResponseMessage.warning("操作失败");
+		}
+		DataMonitorSubmodule submodule = submodules.get(0);
 		if (!submoduleService.deleteByPrimaryKey(submodule.getId())) {
 			return ResponseMessage.warning("操作失败"); 
 		}
